UI-FlaskReact-AppEngineFlexVM/FlaskApiProvider/requirements.txt
absl-py==0.2.2
astor==0.6.2
bleach==2.1.3
CacheControl==0.12.5
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
dill==0.2.8.2
dnspython==1.15.0
dnspython3==1.15.0
firebase-admin==2.11.0
Flask==1.0.2
Flask-Cors==3.0.6
flask_mail==0.9.1
future==0.16.0
gapic-google-cloud-datastore-v1==0.90.4
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
gast==0.2.0
google-api-core==1.2.1
google-api-python-client==1.7.3
google-auth==1.5.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
google-cloud==0.33.1
google-cloud-bigquery==1.3.0
google-cloud-bigquery-datatransfer==0.1.1
google-cloud-bigtable==0.29.0
google-cloud-container==0.1.1
google-cloud-core==0.28.1
google-cloud-datastore==1.6.0
google-cloud-dns==0.29.0
google-cloud-error-reporting==0.30.0
google-cloud-firestore==0.29.0
google-cloud-language==1.0.2
google-cloud-logging==1.6.0
google-cloud-monitoring==0.30.0
google-cloud-pubsub==0.35.4
google-cloud-resource-manager==0.28.1
google-cloud-runtimeconfig==0.28.1
google-cloud-spanner==1.4.0
google-cloud-speech==0.34.0
google-cloud-storage==1.10.0
google-cloud-trace==0.19.0
google-cloud-translate==1.3.1
google-cloud-videointelligence==1.2.0
google-cloud-vision==0.32.0
google-gax==0.16.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpc-google-iam-v1==0.11.4
grpcio==1.12.1
gunicorn==19.8.1
html5lib==1.0.1
httplib2==0.11.3
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
msgpack==0.5.6
nose==1.3.7
numpy==1.14.5
oauth2client==4.1.2
oauthlib==2.1.0
passlib==1.7.1
ply==3.11
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
protobuf==3.6.0
psutil==5.4.6
pyasn1==0.4.3
pyasn1-modules==0.2.1
pyIsEmail==1.3.1
PySocks==1.6.8
pytz==2018.4
requests==2.20.0
requests-oauthlib==1.0.0
rsa==3.4.2
six==1.11.0
tensorboard==1.8.0
tensorflow==1.8.0
tensorflow-tensorboard==1.5.1
termcolor==1.1.0
tweepy==3.6.0
uritemplate==3.0.0
urllib3==1.23
virtualenv==16.0.0
webencodings==0.5.1
Werkzeug==0.14.1
